requirements.txt
discord
python-dotenv
jishaku
pytz
pydub
EZPaginator